# tests/contracts/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ComplianceIQ Core API",
    "description": "\nThe authoritative Core Platform & Data Service for ComplianceIQ.\n\n**Core owns the truth.** Findings, resources, scans and compliance scores\nare produced and owned here. The AI Service reasons *over* this data via\nthis API and never becomes its owner; the dashboard presents it and must\nkeep working when the AI Service is unavailable.\n\n### Authentication\nEvery `/api/v1` endpoint requires `Authorization: Bearer <JWT>` (RS256).\nThe **tenant is taken from the verified token** \u2014 there is no `tenant_id`\nparameter anywhere in this API, and one cannot be supplied.\n\n### Correlation\nSend `X-Correlation-ID` and it is preserved and echoed; omit it and one\nis generated. It appears on every response, including errors.\n\n### Errors\nEvery error \u2014 including validation and unexpected failures \u2014 returns the\nsame envelope: `{\"error\": {\"code\", \"message\", \"correlation_id\", \"details\"}}`.\nBranch on `code`; it is stable. `message` is for humans.\n\n### Versioning\nBreaking changes require a new prefix (`/api/v2`). Additive fields may\nappear in `v1` responses, so clients must ignore unknown fields.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/scans/{scan_id}/attack-paths": {
      "get": {
        "tags": [
          "attack-paths"
        ],
        "summary": "Attack paths discovered by one scan",
        "description": "Every attack path from one scan, **highest risk first**, with a severity summary alongside.\n\nThe tenant comes from the verified JWT; there is no `tenant_id` parameter. A scan belonging to another tenant returns an empty list rather than an error, for the same reason the single-path endpoint returns 404.\n\nEach path carries its full ordered chain and its `evidence.score_factors` breakdown \u2014 the score is only useful if it can be defended.",
        "operationId": "list_attack_paths_for_scan_api_v1_scans__scan_id__attack_paths_get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 512,
              "description": "The scan key.",
              "title": "Scan Id"
            },
            "description": "The scan key."
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(critical|high|medium|low)$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only paths at this severity.",
              "title": "Severity"
            },
            "description": "Only paths at this severity."
          },
          {
            "name": "scenario",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only paths from this analyzer scenario.",
              "title": "Scenario"
            },
            "description": "Only paths from this analyzer scenario."
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(unknown|low|medium|high)$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only paths at or above this confidence. `medium` excludes the low-confidence paths a conditioned IAM grant produces.",
              "title": "Min Confidence"
            },
            "description": "Only paths at or above this confidence. `medium` excludes the low-confidence paths a conditioned IAM grant produces."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttackPathListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or badly signed token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but lacking the 'reader' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter or enum value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/attack-paths/{attack_path_id}": {
      "get": {
        "tags": [
          "attack-paths"
        ],
        "summary": "One attack path",
        "description": "The complete chain, its evidence and its scoring breakdown.\n\n`nodes` and `edges` are **ordered** \u2014 a path whose hops reorder is a different path.\n\n`contributing_finding_ids` links back to the findings on resources along the chain, so a responder can pivot between the composite risk and the individual misconfigurations that create it.",
        "operationId": "get_attack_path_api_v1_attack_paths__attack_path_id__get",
        "parameters": [
          {
            "name": "attack_path_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 1024,
              "description": "The attack path id.",
              "title": "Attack Path Id"
            },
            "description": "The attack path id."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttackPathResource"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, expired or badly signed token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but lacking the 'reader' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter or enum value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such attack path **in this tenant**. Returned identically whether the path does not exist or belongs to another tenant, so the response cannot be used to probe for other tenants' data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AttackPathEdge": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "relationship": {
            "type": "string",
            "title": "Relationship"
          },
          "blocked": {
            "type": "boolean",
            "title": "Blocked",
            "default": false
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence",
            "default": {}
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "target",
          "relationship",
          "confidence"
        ],
        "title": "AttackPathEdge",
        "description": "One relationship traversed."
      },
      "AttackPathListResponse": {
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/AttackPathSummary"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AttackPathResource"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "AttackPathListResponse",
        "description": "Paths plus their severity summary, in one round trip.\n\nThe dashboard needs both to render its landing screen, and two\nendpoints would guarantee they eventually disagree."
      },
      "AttackPathNode": {
        "properties": {
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "resource_id",
          "resource_type",
          "confidence",
          "kind"
        ],
        "title": "AttackPathNode",
        "description": "One hop in an attack path."
      },
      "AttackPathResource": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "scan_key": {
            "type": "string",
            "title": "Scan Key"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/AttackPathNode"
            },
            "type": "array",
            "title": "Nodes",
            "default": []
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/AttackPathEdge"
            },
            "type": "array",
            "title": "Edges",
            "default": []
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence",
            "default": {}
          },
          "contributing_finding_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Contributing Finding Ids",
            "default": []
          },
          "algorithm_version": {
            "type": "string",
            "title": "Algorithm Version"
          },
          "scoring_model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoring Model Version"
          },
          "fingerprint": {
            "type": "string",
            "title": "Fingerprint"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "scan_key",
          "scenario",
          "severity",
          "risk_score",
          "confidence",
          "source",
          "target",
          "algorithm_version",
          "fingerprint",
          "created_at"
        ],
        "title": "AttackPathResource",
        "description": "An attack path as the API exposes it (STEP 5).\n\nCarries the full chain and its scoring breakdown, because the number\nis only useful if it can be defended. ``evidence.score_factors``\nnames every contribution and penalty that produced ``risk_score``.\n\n``nodes`` and ``edges`` are ORDERED \u2014 a path whose hops reorder is a\ndifferent path, and the UI renders them as a sequence."
      },
      "AttackPathSummary": {
        "properties": {
          "critical": {
            "type": "integer",
            "title": "Critical",
            "default": 0
          },
          "high": {
            "type": "integer",
            "title": "High",
            "default": 0
          },
          "medium": {
            "type": "integer",
            "title": "Medium",
            "default": 0
          },
          "low": {
            "type": "integer",
            "title": "Low",
            "default": 0
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AttackPathSummary",
        "description": "Counts by severity \u2014 the dashboard's landing view (\u00a720)."
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code. Branch on this."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation. May be reworded; do not parse."
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id",
            "description": "Echoes X-Correlation-ID; quote it in bug reports."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "correlation_id"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "The response body of EVERY error this API returns."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "RS256 JWT issued by complianceiq-core. Required claims: sub, tenant_id, roles, iss=complianceiq-core, aud=complianceiq, exp. Verify against /.well-known/jwks.json."
      }
    }
  }
}